Update urls.py
Add the include() function allows referencing other URLconfs, , it chops off whatever part of the URL matched up to that point and sends the remaining string to the included URLconf for further processing.

The idea behind include() is to make it easy to plug-and-play URLs. Since polls are in their own URLconf (polls/urls.py), they can be placed under “/polls/”, or under “/fun_polls/”, or under “/content/polls/”, or any other path root, and the app will still work.

Details:
<https://docs.djangoproject.com/en/2.0/intro/tutorial01/>